refactor: refactor profile head
