+++ packages/author-head/package.json
@@ -69,6 +69,7 @@
     "react-test-renderer": "16.3.1"
   },
   "dependencies": {
+    "@times-components/gradient": "0.10.7",
     "@times-components/icons": "0.7.7",
     "@times-components/image": "1.17.5",
     "@times-components/link": "0.20.7",
