+++ packages/author-profile/src/author-profile-pagination.js
@@ -1,23 +1,7 @@
 import React from "react";
-import { StyleSheet, View } from "react-native";
+import { View } from "react-native";
 import Pagination from "@times-components/pagination";
-import { spacing } from "@times-components/styleguide";
-
-const styles = StyleSheet.create({
-  container: {
-    alignItems: "stretch",
-    flexDirection: "row",
-    justifyContent: "center"
-  },
-  itemContainer: {
-    paddingLeft: spacing(2),
-    paddingRight: spacing(2)
-  },
-  spacing: {
-    flex: 1,
-    maxWidth: 800
-  }
-});
+import styles from "../styles";
 
 const AuthorProfilePagination = props => (
   <View style={styles.container}>
@@ -28,4 +12,5 @@ const AuthorProfilePagination = props => (
 );
 
 AuthorProfilePagination.propTypes = Pagination.propTypes;
+
 export default AuthorProfilePagination;
